requirements.txt
asgiref==3.8.1
attrs==25.3.0
Automat==24.8.1
azure-ai-ml==1.26.0
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-storage-blob==12.25.0
azure-storage-file-datalake==12.19.0
azure-storage-file-share==12.21.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
filelock==3.18.0
fixedint==0.1.6
Flask==3.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
isodate==0.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
parsel==1.10.0
Protego==0.4.0
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydash==8.0.5
PyDispatcher==2.0.7
PyJWT==2.10.1
pymongo==4.11.3
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
strictyaml==1.7.3
threadpoolctl==3.6.0
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
w3lib==2.3.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zope.interface==7.2
